# Speed up PNG and improve JPEG quality