15x faster saving PNG and 2x smaller JPEGs.

PNG
Seconds | Old | New | Speed | Size
| :---: | :---: | :---: | :---: | :---: |
1080p | 7.77 | 0.53 | 14.66x | +10%
1440p | 13.67 | 0.91 | 15.02x | +10%
2160p | 31.21 | 1.92 | 16.26x | +14%
4320p | 126.4 | 7.63 | 16.56x | +20%

JPEG
4K | Seconds | Size (MB) | SSIMU2 (Quality)
| :---: | :---: | :---: | :---: |
Old | 0.26 | 3.24 | 87.19
New | 0.46 | 1.31 | 86.98